execCommand('print', false, null) & IE meta tag
Utilize execCommand('print', false, null) in browsers that support it to
avoid issues with Internet Explorer. I've been using this method in a
production site with thousands of daily active users since February and
haven't had any reported issues. Close #43. Works in IE8-11, and falls
back to the previous method if the browser doesn't support it.

Also, in IE9, the doctype wasn't being transferred correctly to the
iframe, and was causing display issues, as CSS child selectors only work
if the doctype is set properly. Adding a compatibility meta tag for IE
resolved this issue.